+++ drizzle/0015_new_pet_avengers.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `cassettes` ADD `erstauflage` integer DEFAULT false NOT NULL;